fix: use row class rules instead of row classes
